[*] Move files to more descriptive locations
